fix: keep feature flags client-safe
